feat: add text splitter
