+++ llmchain/loaders/tests/it/directory/directory_splitter.rs
@@ -18,7 +18,7 @@ use anyhow::Result;
 use goldenfile::Mint;
 use llmchain_loaders::directory::DirectoryLoader;
 use llmchain_loaders::document::DocumentLoader;
-use llmchain_loaders::document_splitter::TextSplitter;
+use llmchain_loaders::document_splitter::DocumentSplitter;
 use llmchain_loaders::markdown::MarkdownLoader;
 use llmchain_loaders::markdown::MarkdownSplitter;
 use opendal::services::Fs;
@@ -56,7 +56,7 @@ fn test_directory_splitter_default() -> Result<()> {
             "part={}, len={}, chunk_size={}, path={}",
             i,
             doc.content.len(),
-            markdown_splitter.settings().splitter_chunk_size,
+            markdown_splitter.splitter_chunk_size,
             doc.meta.path
         )?;
         writeln!(
